Update Firebase-Write-Permission-Exploit.py
Updating this as there was an error causing an issue with exploit to fail when the the rules are setup for false for read but true for write. making it useless when a user can't read the database but still write. The issue started to appear when a Contributor added his code...